fix: make horizontal chart scaling smooth by addressing precision loss due to rounding

+++ src/main/kotlin/io/redgreen/timelapse/geometry/LineFunctions.kt
@@ -1,4 +1,4 @@
 package io.redgreen.timelapse.geometry
 
-fun calculateSlope(x1: Int, y1: Int, x2: Int, y2: Int): Double =
-  (y2 - y1).toDouble() / (x2 - x1)
+fun calculateSlope(x1: Double, y1: Int, x2: Double, y2: Int): Double =
+  (y2 - y1) / (x2 - x1)
